requirements.txt
langchain
langchain-ollama
langchain-huggingface
langchain-community
chromadb
sentence-transformers
python-dotenv
flask 
sentence-transformers 
faiss-cpu